+++ changelog.md
@@ -6,6 +6,7 @@
 
 - [#2950](https://github.com/ignite/cli/pull/2950) Add `--genesis-config` flag option to `ignite network chain publish`
 - [#2955](https://github.com/ignite/cli/pull/2955/) Add `ignite network request add-account` command.
+- [#2877](https://github.com/ignite/cli/pull/2877) Plugin system
 - [#2995](https://github.com/ignite/cli/pull/2995/) Add `ignite network request remove-validator` command.
 - [#2999](https://github.com/ignite/cli/pull/2999/) Add `ignite network request remove-account` command.